completed the delete function for contact forms

+++ ChucksUsedDealership/Controllers/ContactFormController.cs
@@ -1,5 +1,6 @@
 ﻿using ChucksUsedDealership.Data;
 using ChucksUsedDealership.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ChucksUsedDealership.Controllers
@@ -35,14 +36,17 @@ namespace ChucksUsedDealership.Controllers
             return View("Index", model);
         }
 
+        [Authorize(Roles ="Admin, Authorized")]
+        [HttpGet]
         public IActionResult ContactFormList()
         {
             var contactForms = _context.ContactForms.ToList();
             return View(contactForms);
         }
 
+        [Authorize(Roles = "Admin, Authorized")]
         [HttpGet]
-        public async Task<IActionResult> ContactFormDelete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
             if (id == null)
             {
@@ -57,14 +61,21 @@ namespace ChucksUsedDealership.Controllers
             return View(contactForm);
         }
 
+        [Authorize(Roles = "Admin, Authorized")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> ContactFormDeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int ContactFormId)
         {
-            var contactForm = await _context.ContactForms.FindAsync(id);
+            var contactForm = await _context.ContactForms.FindAsync(ContactFormId);
+            if (contactForm == null)
+            {
+                return NotFound();
+            }
             _context.ContactForms.Remove(contactForm);
             await _context.SaveChangesAsync();
+            TempData["SuccessMessage"] = "Contact Form deleted successfully!";
             return RedirectToAction(nameof(ContactFormList));
         }
+
     }
 }
